Implement updated tree viewing with tests
This code reformats the way the tree is displayed, now showing a structure that visually resembles a tree. It also includes comprehensive tests for this new display implementation. These changes should improve readability and verify that the tree is correctly representing the data structure.
